requirements.txt
blinker==1.7.0
cffi==1.16.0
chardet==5.2.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
DateTime==5.4
dlib @ file:///D:/without-front-end/dlib-19.22.99-cp310-cp310-win_amd64.whl
et-xmlfile==1.1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==3.0.0
Flask-Assets==2.1.0
Flask-Cors==4.0.0
Flask-SSLify==0.1.5
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.4
MarkupSafe==2.1.3
nmpy==0.0.1
numpy==1.26.2
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pycparser==2.21
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-pptx==0.6.23
pytz==2023.3.post1
reportlab==4.0.8
six==1.16.0
typing_extensions==4.9.0
tzdata==2023.4
webassets==2.0
Werkzeug==3.0.1
XlsxWriter==3.1.9
zope.interface==6.1